# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VmConfig {
  /** Path or URL to disk image */
  diskImage: string;
  /** RAM in MB */
  ram: number;
  /** Number of CPU cores */
  cpus: number;
  audioEnabled?: boolean;
  /** QEMU audio device type */
  audioDevice?: string;
  /** QEMU machine type (pc, q35) */
  pcMode?: string;
  networkEnabled?: boolean;
  /** VGA adapter type (std, virtio, cirrus) */
  vgaType?: string;
  /** Boot order (c=HDD, d=CD, n=Network) */
  bootOrder?: string;
}

export interface VmStatus {
  running: boolean;
  /** @nullable */
  pid?: number | null;
  /** @nullable */
  vncPort?: number | null;
  /** @nullable */
  audioWsPort?: number | null;
  /** @nullable */
  uptime?: number | null;
  config?: VmConfig;
}

export interface DiskImage {
  name: string;
  path: string;
  /** @nullable */
  sizeBytes?: number | null;
  isDefault?: boolean;
}

export interface Snapshot {
  id: string;
  name: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}

export interface SnapshotInput {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface KeyInput {
  /** Key combo string e.g. "ctrl-alt-delete", "win-r", "f1" */
  combo: string;
}

export interface TextInput {
  text: string;
}

export interface TwitchConfig {
  channel: string;
  enabled: boolean;
  commandPrefix?: string;
  /** Empty = all users allowed */
  allowedUsers?: string[];
}

export interface TwitchCommandEntry {
  id: string;
  username: string;
  rawCommand: string;
  parsedCommands: string[];
  executedAt: string;
  /** executed | queued | error */
  status: string;
}

export interface OperationResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}