# lib/api-zod/src/generated/types/diskImage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */

export interface DiskImage {
  name: string;
  path: string;
  /** @nullable */
  sizeBytes?: number | null;
  isDefault?: boolean;
}