# lib/api-zod/src/generated/types/keyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */

export interface KeyInput {
  /** Key combo string e.g. "ctrl-alt-delete", "win-r", "f1" */
  combo: string;
}