# lib/api-zod/src/generated/types/operationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */

export interface OperationResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}