# lib/api-zod/src/generated/types/snapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */

export interface Snapshot {
  id: string;
  name: string;
  createdAt: Date;
  /** @nullable */
  description?: string | null;
}