# lib/api-zod/src/generated/types/twitchConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */

export interface TwitchConfig {
  channel: string;
  enabled: boolean;
  commandPrefix?: string;
  /** Empty = all users allowed */
  allowedUsers?: string[];
}