# lib/api-zod/src/generated/types/vmConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */

export interface VmConfig {
  /** Path or URL to disk image */
  diskImage: string;
  /** RAM in MB */
  ram: number;
  /** Number of CPU cores */
  cpus: number;
  audioEnabled?: boolean;
  /** QEMU audio device type */
  audioDevice?: string;
  /** QEMU machine type (pc, q35) */
  pcMode?: string;
  networkEnabled?: boolean;
  /** VGA adapter type (std, virtio, cirrus) */
  vgaType?: string;
  /** Boot order (c=HDD, d=CD, n=Network) */
  bootOrder?: string;
}