# lib/api-zod/src/generated/types/vmStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QEMU Web Emulator API
 * OpenAPI spec version: 0.1.0
 */
import type { VmConfig } from './vmConfig';

export interface VmStatus {
  running: boolean;
  /** @nullable */
  pid?: number | null;
  /** @nullable */
  vncPort?: number | null;
  /** @nullable */
  audioWsPort?: number | null;
  /** @nullable */
  uptime?: number | null;
  config?: VmConfig;
}